+++ packages/numbers/index.js
@@ -1,9 +1,23 @@
 // import packages
+const addLeadingZero = require('../../utils/numbers/addLeadingZero')
+const addTrailingZeros = require('../../utils/numbers/addTrailingZeros')
+const getAverage = require('../../utils/numbers/getAverage')
+const getDiff = require('../../utils/numbers/getDiff')
+const getRandomInRange = require('../../utils/numbers/getRandomInRange')
+const getSum = require('../../utils/numbers/getSum')
 const isEven = require('../../utils/numbers/isEven')
+const roundTo = require('../../utils/numbers/roundTo')
 const toReadable = require('../../utils/numbers/toReadable')
 
 // export packages
 module.exports = {
+	addLeadingZero,
+	addTrailingZeros,
+	getAverage,
+	getDiff,
+	getRandomInRange,
+	getSum,
 	isEven,
+	roundTo,
 	toReadable,
 }