+++ utils/numbers/addLeadingZero.js
@@ -1,0 +1,2 @@
+// add leading zero if not existent
+module.exports = (value) => (value < 10 ? `0${value}` : value).toString()