+++ utils/numbers/getDiff.js
@@ -1,0 +1,2 @@
+// get difference of two values
+module.exports = (a, b) => a - b